Fixed Build Error [Test]
As mentioned in private already, you have to cast the UINT_MAX to a uint. While a long might've worked for you, that's actually just luck. To improve memory usage and return to best-practices I would recommend to use a uint64_t instead of a long. I already proposed that in private.